refactor: add attribute service
